+++ src/main/java/com/pedidos/repository/PedidoRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface PedidoRepository extends MongoRepository<Pedido, String> {
-    Optional<Pedido> findFirstByCodigoPedido (int codigoPedido);
-    List<Pedido> findByCodigoCliente(int codigoCliente);
+    Optional<Pedido> findFirstByCodigoPedido (long codigoPedido);
+    List<Pedido> findByCodigoCliente(long codigoCliente);
 }
